# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Handassa",
        "version": "1.0.0"
    },
    "paths": {
        "/api/check_username": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Check Username",
                "operationId": "checkUsername",
                "requestBody": {
                    "description": "Check Username",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Person1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/check_mobile_number": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Check Mobile Number",
                "operationId": "checkMobileNumber",
                "requestBody": {
                    "description": "Check Mobile Number",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mobile_number"
                                ],
                                "properties": {
                                    "mobile_number": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "+921234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Create Account",
                "operationId": "register",
                "requestBody": {
                    "description": "Register",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "image to upload",
                                        "type": "file",
                                        "format": "file"
                                    },
                                    "cr copy": {
                                        "description": "cr copy to upload for company",
                                        "type": "file",
                                        "format": "file"
                                    },
                                    "id copy": {
                                        "description": "id copy to upload for freelancer",
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "required": [
                                    "role",
                                    "type",
                                    "username",
                                    "name",
                                    "mobile_number",
                                    "location"
                                ],
                                "properties": {
                                    "role": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "client"
                                    },
                                    "type": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "person"
                                    },
                                    "username": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "person1"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "My Name"
                                    },
                                    "mobile_number": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "+921234567890"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "+921234567890 for freelancer company"
                                    },
                                    "location": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Al Doha"
                                    },
                                    "services": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "[1,2] for professional"
                                    },
                                    "categories": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "[1,2] for professional"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "operationId": "login",
                "requestBody": {
                    "description": "Login",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mobile_number"
                                ],
                                "properties": {
                                    "mobile_number": {
                                        "type": "string",
                                        "format": "mobile_number",
                                        "example": "+921234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/locations": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Locations",
                "operationId": "Locations",
                "responses": {
                    "default": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/services": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Services",
                "operationId": "Services",
                "responses": {
                    "default": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/categories": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Categories",
                "operationId": "Categories",
                "responses": {
                    "default": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/upload_image": {
            "post": {
                "tags": [
                    "Upload Image"
                ],
                "summary": "Upload Image",
                "operationId": "uploadImage",
                "requestBody": {
                    "description": "Upload Image",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "image to upload",
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": " All Posts",
                "operationId": "allPosts",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "new"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/client/save/post": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Save Post",
                "operationId": "savePost",
                "requestBody": {
                    "description": "Save Post",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "service_id",
                                    "category_id",
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "service_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "description"
                                    },
                                    "job": {
                                        "type": "string",
                                        "format": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/professional/post/proposal": {
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Post Proposal",
                "operationId": "postProposal",
                "requestBody": {
                    "description": "Post Proposal",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "post_id",
                                    "price",
                                    "description"
                                ],
                                "properties": {
                                    "post_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    },
                                    "price": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "20.22"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}